requirements.txt
Django==4.2.0
python-dotenv==0.20.0
psycopg2==2.9.5
djangorestframework==3.14.0
django-debug-toolbar==4.1.0
redis==5.0.0